fix: refactoring LoadingTracker Component
