+++ pages/collection/[collection_id].js
@@ -32,6 +32,7 @@ import CollectionSearch from 'components/Collection/CollectionSearch'
 import { SHOW_TX_HASH_LINK } from 'constants/common'
 import DailyTracker from 'components/LineChart/DailyTracker'
 import Link from 'next/link'
+import LoadingTracker from 'components/Common/LoadingTracker'
 
 const LIMIT = 12
 const LIMIT_ACTIVITY = 20
@@ -1037,42 +1038,4 @@ export async function getServerSideProps({ params }) {
 			serverQuery: params,
 		},
 	}
-}
-
-const LoadingTracker = () => {
-	return (
-		<div className="md:grid grid-cols-2">
-			<LoadingTrackerIcon />
-			<LoadingTrackerIcon />
-			<LoadingTrackerIcon />
-			<LoadingTrackerIcon />
-		</div>
-	)
-}
-
-const LoadingTrackerIcon = () => {
-	return (
-		<div className="flex items-center justify-center w-full h-80">
-			<svg
-				className="animate-spin m h-5 w-5 text-white"
-				xmlns="http://www.w3.org/2000/svg"
-				fill="none"
-				viewBox="0 0 24 24"
-			>
-				<circle
-					className="opacity-25"
-					cx="12"
-					cy="12"
-					r="10"
-					stroke="currentColor"
-					strokeWidth="4"
-				></circle>
-				<path
-					className="opacity-75"
-					fill="currentColor"
-					d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-				></path>
-			</svg>
-		</div>
-	)
 }
